Fix MSVC size conversion warnings

+++ src/findfile.c
@@ -1166,7 +1166,8 @@ vim_findfile(void *search_ctx_arg)
 			    if (*suf == NUL)
 				break;
 			    file_path.length = len + copy_option_part(&suf,
-				  file_path.string + len, MAXPATHL - len, ",");
+				  file_path.string + len,
+				  (int)(MAXPATHL - len), ",");
 			}
 		    }
 		}
